Implement Order and OrderItem repository functions
This commit introduces the functions to create Order and OrderItem objects in the database using the Prisma ORM. It handles the creation of each OrderItem associated with an Order and ensures that all created data are properly persisted in the database. In case of an error during this process, it handles the exception and returns an error message.