Add payment handling in Order Form
The update replaces the CashPayment and CardPayment components in the Create Order Modal form. Payments are now being added and removed dynamically depending on the selected payment method. Specifically, when the payment method changes, the previous method payment information gets removed and a new one for the currently selected method is added.