Implement order creation in payment modal
Added order creation logic in the payment modal, along with success and error handling toast notifications. This includes the use of a new action to create orders and reset the form in case the order creation is successful. In case of failure, an error message is displayed using a toast notification.