Add payment details to order data component
This commit introduces the display of various payment methods within the order data component. The changes involve mapping each payment method to a human-readable string in the front-end and adding that payment method to the database retrieval in the back-end. Hence, each purchase now has visible details of the payment method used.